malabarhouse/requirements.txt
app-version==0.2.1
Django==1.8.3
django-appconf==1.0.1
django-inspectional-registration==0.5.0
django-multiselectfield==0.1.3
six==1.9.0
tolerance==0.1.2
